Refactor mobDetectTest.js to remove unused code and improve entity tracking
