+++ features/Diana/DianaMobDetect.js
@@ -1,7 +1,7 @@
 import settings from "../../settings";
 import { registerWhen } from "../../utils/variables";
 import { getWorld } from "../../utils/world";
-import { mythosMobHpOverlay } from "./../guis/DianaGuis";
+import { mythosMobHpOverlay } from "../guis/DianaGuis";
 import { checkDiana } from "../../utils/checkDiana";
 import RenderLibV2 from "../../../RenderLibV2";
 import { data } from "../../utils/variables";
@@ -17,76 +17,141 @@ register("command", () => {
     }
 }).setName("sboinqmsgtest");
 
-let Mobs = [];
-let inqs = [];
-registerWhen(register("step", () => {
-    if (checkDiana()) {
-        World.getAllEntitiesOfType(net.minecraft.entity.item.EntityArmorStand).forEach((mob) => {
-            let name = mob.getName();   
-            if (name != "Armor Stand") {
-                if (Mobs.filter((e) => e.getUUID() === mob.getUUID()).length === 0) {
-                    if ((name.includes("Exalted") || name.includes("Stalwart")) && !name.split(" ")[2].startsWith("0")) { //  || name.includes("Graveyard Zombie")
-                        Mobs.push(mob);
+let TrackedInqs = {};
+let TrackedMobs = {};
+
+function trackEntity(uuid, mob, dictionary, entityTypeCheck, entityClasses, useOldTracking = false) {
+    if (entityTypeCheck(mob.getName())) {
+        if (!dictionary[uuid]) {
+            dictionary[uuid] = {
+                entity: mob,
+                x: mob.getX(),
+                y: mob.getY(),
+                z: mob.getZ(),
+                trackedEntity: null,
+                lastSeen: Date.now(),
+            };
+        } else {
+            let data = dictionary[uuid];
+            data.x = mob.getX();
+            data.y = mob.getY();
+            data.z = mob.getZ();
+            data.lastSeen = Date.now();
+
+            if (useOldTracking) {
+                return;
+            }
+            
+
+            if (!data.trackedEntity) {
+                for (let entityClass of entityClasses) {
+                    let foundEntity = World.getAllEntities()
+                        .filter(entity => entity instanceof entityClass)
+                        .find(entity => 
+                            entity.distanceTo(mob) < 3 && 
+                            Math.abs(entity.getX() - mob.getX()) <= 1 && 
+                            Math.abs(entity.getZ() - mob.getZ()) <= 1
+                        );
+
+                    if (foundEntity) {
+                        data.trackedEntity = foundEntity;
+                        break;
                     }
-                    
                 }
             }
-        });
-        Mobs = Mobs.filter((e) => !e.getEntity()["field_70128_L"]);
-        inqs = Mobs.filter((mob) => mob.getName().includes("Inquisitor")); // || mob.getName().includes("Graveyard Zombie")
-    }
-}).setFps(1), () => settings.mythosMobHp || settings.inqHighlight || settings.inqCircle && getWorld() === "Hub");
 
-let names = [];
+            if (data.trackedEntity) {
+                data.x = data.trackedEntity.getX();
+                data.y = data.trackedEntity.getY();
+                data.z = data.trackedEntity.getZ();
+            }
+        }
+    }
+}
+
+function updateTrackedEntities(trackedEntities, seenUUIDs) {
+    for (let uuid in trackedEntities) {
+        let mob = trackedEntities[uuid];
+        if (!seenUUIDs.has(uuid)) {
+            if (mob.trackedEntity) {
+                if (mob.trackedEntity.getEntity()["field_70128_L"]) {
+                    delete trackedEntities[uuid];
+                    continue;
+                }
+                mob.x = mob.trackedEntity.getX();
+                mob.y = mob.trackedEntity.getY();
+                mob.z = mob.trackedEntity.getZ();
+                mob.lastSeen = Date.now();
+            } else if (Date.now() - mob.lastSeen > 500) {
+                delete trackedEntities[uuid];
+            }
+        }
+    }
+}
+
+const dianaEntityList = [
+    net.minecraft.entity.monster.EntityZombie,
+    net.minecraft.entity.passive.EntityIronGolem,
+    net.minecraft.entity.passive.EntityOcelot
+];
+
 registerWhen(register("step", () => {
     if (checkDiana()) {
-        names = [];
-        Mobs.forEach((nameTag) => {
-            names.push(nameTag.getName());
+        let seenUUIDs = new Set();
+        
+        World.getAllEntitiesOfType(net.minecraft.entity.item.EntityArmorStand).forEach((armorStand) => {
+            let uuid = armorStand.getUUID();
+            seenUUIDs.add(uuid);
+            
+            trackEntity(uuid, armorStand, TrackedInqs, (name) => name.includes("Inquisitor"), dianaEntityList);  
+            trackEntity(uuid, armorStand, TrackedMobs, (name) => (name.includes("Exalted") || 
+            name.includes("Stalwart")) && !name.split(" ")[2].startsWith("0"), dianaEntityList, true);
         });
-        mythosMobHpOverlay(names);
+        
+        updateTrackedEntities(TrackedInqs, seenUUIDs);
+        updateTrackedEntities(TrackedMobs, seenUUIDs);
     }
-}).setFps(6), () => settings.mythosMobHp && getWorld() === "Hub");
+}).setFps(6), () => settings.mythosMobHp || settings.inqHighlight || settings.inqCircle && getWorld() === "Hub");
 
-   
+let inqs = [];
 export const inqHighlightRegister = register("renderWorld", () => {
-    inqs.forEach((mob) => {
+    inqs = [];
+    Object.values(TrackedInqs).forEach((mob) => {
+        let yOffset = mob.trackedEntity ? 0 : -2.05;
         let red = settings.inqColor.getRed() / 255;
         let green = settings.inqColor.getGreen() / 255;
         let blue = settings.inqColor.getBlue() / 255;
         let alpha = settings.inqColor.getAlpha() / 255;
+
         if (settings.inqHighlight) {
-            RenderLibV2.drawEspBoxV2(mob.x, mob.y - 2.05, mob.z, 1, 2, 1, red, green, blue, alpha, false)   
+            RenderLibV2.drawEspBoxV2(mob.x, mob.y + yOffset, mob.z, 1, 2, 1, red, green, blue, alpha, false);
         }
         if (settings.inqCircle) {
-            let hight = 0.6;
-            let y = mob.y - 2.05;
+            let height = 0.6;
+            let y = mob.y + yOffset;
             if (settings.inqCylinder) {
-                hight = 50;
+                height = 50;
                 y = 50;
             }
-            RenderLibV2.drawCyl(mob.x, y, mob.z, 30, 30, hight, 120, 1, 0, 90, 90, red, green, blue, alpha, false, false)
+            RenderLibV2.drawCyl(mob.x, y, mob.z, 30, 30, height, 120, 1, 0, 90, 90, red, green, blue, alpha, false, false);
         }
     });
 });
 inqHighlightRegister.unregister();
 
-// register("tick", () => {
-//     // check if player is inside circle
-//     // player x = Player.getX()
-//     // player z = Player.getZ()
-//     let inCircle = false;
-//     inqs.forEach((mob) => {
-//         let distance = Math.sqrt((mob.x - Player.getX()) ** 2 + (mob.z - Player.getZ()) ** 2);
-//         if (distance < 30) {
-//             inCircle = true;
-//         }
-//     });
-//     ChatLib.chat("circle: " + inCircle);
-// });
+let names = [];
+registerWhen(register("step", () => {
+    if (checkDiana()) {
+        names = [];
+        Object.values(TrackedMobs).forEach((mob) => {
+            names.push(mob.entity.getName());
+        });
+        mythosMobHpOverlay(names);
+    }
+}).setFps(6), () => settings.mythosMobHp && getWorld() === "Hub");
 
 register("worldUnload", () => {
-    Mobs = [];
-    inqs = [];
+    TrackedInqs = {};
+    TrackedMobs = {};
     names = [];
 });
